
\name{Cuadrant}
\alias{Cuadrant}
\title{
Cuadrant 
}
\description{
Esta funci칩n realiza el test de aleatoriedad por el algoritmo de recorte de cuadrantes
}
\usage{
Cuadrant(Carpeta, graf = FALSE)
}
\arguments{
  \item{Carpeta}{
par치metro de entrada, en el cual se coloca el objeto a ser explorado, el objeto generalmente es un conjunto de datos por ejemplo: un objeto resultante de la funci칩n trimExt.
}
  \item{graf}{
se indica si se desea obtener las gr치ficas de los cuadrantes (TRUE), Default=(FALSE)
}
}

\references{
Diggle, P.J. Statistical analysis of spatial point patterns. Academic Press, 2003. \cr
Stoyan, D. and Stoyan, H. (1994) Fractals, random shapes and point fields: methods of geometrical statistics. John Wiley and Sons.
}
\author{
Camilo Alberto Herrera Rozo  \email{ hr.camilo@hrglobalideas.com}
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
