\name{NeuroSpatCo-package}
\alias{NeuroSpatCo-package}
\alias{NeuroSpatCo}
\docType{package}
\title{NeuroSpatCo}
\description{Este paquete es una implementación realizada para la tesis de pregrado de Camilo A Herrera, Estudiante de Estadística de la Universidad del Valle, Colombia. El paquete sirve para la búsqueda de patrones organizacionales en la corteza pre-frontal de los humanos tanto en plano Horizontal somas como en plano vertical minicolumnas.
}
\details{
\tabular{ll}{
Package: \tab NeuroSpatCo\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2011-06-22\cr
License: \tab GPL (>=2) \cr
LazyLoad: \tab yes\cr
}
}
\author{Camilo Alberto Herrera Rozo, Contact: Camilo Alberto Herrera Rozo  \email{ hr.camilo@hrglobalideas.com}
}
\keyword{ package }


