\name{NeuroSpatCo-package}
\alias{NeuroSpatCo-package}
\alias{NeuroSpatCo}
\docType{package}
\title{NeuroSpatCo}
\description{
Este paquete es una implementación realizada para la tesis de pregrado de Camilo A Herrera, Estudiante de Estadistica de la Universidad del Valle - Colombia. El paquete sirve para la busqueda de patrones organizacionales en la corteza prefrontal de los humanos tanto en plano Horizontal somas como en plano Vertical minicolumnas.
}
\details{
\tabular{ll}{
Package: \tab NeuroSpatCo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-04-19\cr
License: \tab GPL (>=2) \cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Camilo Alberto Herrera Rozo
Contact: Camilo Alberto Herrera Rozo  \email{ hr.camilo@hrglobalideas.com}
}
\references{

}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
