\name{read.folder}
\alias{read.folder}
\title{
Read.folder
}
\description{
lee carpetas enteras de datos en archivos .txt con el formato establecido (centroidex,centroidey,area,orientacion,largo,ancho)
}
\usage{
read.folder()
}
\details{
Esta funcion sirve para leer los datos tanto de las imagenes de corte horizontal como las imagenes de corte vertical.
}


\author{
Camilo Alberto Herrera Rozo  \email{ hr.camilo@hrglobalideas.com}
}

\examples{
Carpetas<-read.folder()
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
